1.0.3
Rename:
ViewController -> VC
ViewModel -> VM
ViewOutlet -> VO
