+++ Sources/MVVVR/Detail/DetailRouter.swift
@@ -7,13 +7,14 @@
 import UIKit
 
 extension Detail {
-    @MainActor final class Router {
-        weak var vc: ViewController?
+    @MainActor
+    final class Router {
+        weak var vc: VC?
         
         // MARK: - Public
         
         func toReader(comicId: String, episodeId: String) {
-            let to = Reader.ViewController(comicId: comicId, episodeId: episodeId)
+            let to = Reader.VC(comicId: comicId, episodeId: episodeId)
             to.hidesBottomBarWhenPushed = true
             vc?.navigationController?.show(to, sender: nil)
         }
