+++ Sources/MVVVR/Detail/DetailVO.swift
@@ -1,25 +1,44 @@
 //
-//  EpisodeListVO.swift
+//  DetailVO.swift
 //
-//  Created by Shinren Pan on 2024/6/4.
+//  Created by Shinren Pan on 2024/5/22.
 //
 
 import UIKit
 
-extension EpisodePicker {
-    @MainActor final class ViewOutlet {
+extension Detail {
+    @MainActor
+    final class VO {
         let mainView = UIView(frame: .zero)
+            .setup(\.translatesAutoresizingMaskIntoConstraints, value: false)
+            .setup(\.backgroundColor, value: .white)
+        
+        let header = Header()
             .setup(\.translatesAutoresizingMaskIntoConstraints, value: false)
             .setup(\.backgroundColor, value: .white)
         
         let list = UICollectionView(frame: .zero, collectionViewLayout: makeListLayout())
             .setup(\.translatesAutoresizingMaskIntoConstraints, value: false)
+            .setup(\.refreshControl, value: .init(frame: .zero))
+        
+        let favoriteItem = UIBarButtonItem()
+            .setup(\.image, value: .init(systemName: "star"))
         
         init() {
             addViews()
         }
         
         // MARK: - Public
+        
+        func reloadHeader(comic: DisplayComic) {
+            header.reloadUI(comic: comic)
+        }
+        
+        func reloadFavoriteUI(comic: DisplayComic) {
+            let imgNamed = comic.favorited ? "star.fill" : "star"
+            let image = UIImage(systemName: imgNamed)
+            favoriteItem.image = image
+        }
         
         func scrollListToWatched(indexPath: IndexPath?) {
             guard let indexPath else {
@@ -32,10 +51,15 @@ extension EpisodePicker {
         // MARK: - Add Something
 
         private func addViews() {
+            mainView.addSubview(header)
             mainView.addSubview(list)
 
             NSLayoutConstraint.activate([
-                list.topAnchor.constraint(equalTo: mainView.topAnchor),
+                header.topAnchor.constraint(equalTo: mainView.topAnchor),
+                header.leadingAnchor.constraint(equalTo: mainView.leadingAnchor),
+                header.trailingAnchor.constraint(equalTo: mainView.trailingAnchor),
+
+                list.topAnchor.constraint(equalTo: header.bottomAnchor),
                 list.leadingAnchor.constraint(equalTo: mainView.leadingAnchor),
                 list.trailingAnchor.constraint(equalTo: mainView.trailingAnchor),
                 list.bottomAnchor.constraint(equalTo: mainView.bottomAnchor),
