+++ Sources/MVVVR/EpisodePicker/EpisodePickerRouter.swift
@@ -1,5 +1,5 @@
 //
-//  EpisodeListRouter.swift
+//  EpisodePickerRouter.swift
 //
 //  Created by Shinren Pan on 2024/6/4.
 //
@@ -7,7 +7,8 @@
 import UIKit
 
 extension EpisodePicker {
-    @MainActor final class Router {
-        weak var vc: ViewController?
+    @MainActor
+    final class Router {
+        weak var vc: VC?
     }
 }
