+++ Sources/MVVVR/Favorite/FavoriteRouter.swift
@@ -1,5 +1,5 @@
 //
-//  FavoriteListRouter.swift
+//  FavoriteRouter.swift
 //
 //  Created by Shinren Pan on 2024/5/22.
 //
@@ -7,13 +7,14 @@
 import UIKit
 
 extension Favorite {
-    @MainActor final class Router {
-        weak var vc: ViewController?
+    @MainActor
+    final class Router {
+        weak var vc: VC?
         
         // MARK: - Public
         
         func toDetail(comicId: String) {
-            let to = Detail.ViewController(comicId: comicId)
+            let to = Detail.VC(comicId: comicId)
             to.hidesBottomBarWhenPushed = true
             vc?.navigationController?.show(to, sender: nil)
         }
