+++ Sources/MVVVR/Favorite/FavoriteViews.swift
@@ -3,7 +3,6 @@
 //
 //  Created by Joe Pan on 2024/10/28.
 //
-//
 
 import UIKit
 import SwiftUI
