+++ Sources/MVVVR/History/HistoryRouter.swift
@@ -1,5 +1,5 @@
 //
-//  HistoryListRouter.swift
+//  HistoryRouter.swift
 //
 //  Created by Shinren Pan on 2024/5/23.
 //
@@ -8,12 +8,12 @@ import UIKit
 
 extension History {
     @MainActor final class Router {
-        weak var vc: ViewController?
+        weak var vc: VC?
         
         // MARK: - Public
         
         func toDetail(comicId: String) {
-            let to = Detail.ViewController(comicId: comicId)
+            let to = Detail.VC(comicId: comicId)
             to.hidesBottomBarWhenPushed = true
             vc?.navigationController?.show(to, sender: nil)
         }
