+++ Sources/MVVVR/Search/SearchRouter.swift
@@ -9,12 +9,12 @@ import UIKit
 extension Search {
     @MainActor
     final class Router {
-        weak var vc: ViewController?
+        weak var vc: VC?
         
         // MARK: - Public
         
         func toDetail(comicId: String) {
-            let to = Detail.ViewController(comicId: comicId)
+            let to = Detail.VC(comicId: comicId)
             to.hidesBottomBarWhenPushed = true
             vc?.navigationController?.show(to, sender: nil)
         }
