+++ Sources/MVVVR/Update/UpdateRouter.swift
@@ -1,5 +1,5 @@
 //
-//  UpdateListRouter.swift
+//  UpdateRouter.swift
 //
 //  Created by Shinren Pan on 2024/5/21.
 //
@@ -7,19 +7,20 @@
 import UIKit
 
 extension Update {
-    @MainActor final class Router {
-        weak var vc: ViewController?
+    @MainActor
+    final class Router {
+        weak var vc: VC?
         
         // MARK: - Public
         
         func toDetail(comicId: String) {
-            let to = Detail.ViewController(comicId: comicId)
+            let to = Detail.VC(comicId: comicId)
             to.hidesBottomBarWhenPushed = true
             vc?.navigationController?.show(to, sender: nil)
         }
         
         func toRemoteSearch() {
-            let to = Search.ViewController()
+            let to = Search.VC()
             to.hidesBottomBarWhenPushed = true
             vc?.navigationController?.show(to, sender: nil)
         }
