+++ project.yml
@@ -27,8 +27,8 @@ targets:
       base:
         CODE_SIGN_STYLE: Manual
         INFOPLIST_FILE: Sources/App/Info.plist
-        MARKETING_VERSION: 1.0.2
-        CURRENT_PROJECT_VERSION: 20241105
+        MARKETING_VERSION: 1.0.3
+        CURRENT_PROJECT_VERSION: 20241112
         PRODUCT_BUNDLE_IDENTIFIER: com.shinrenpan.Comic
       configs:
         Debug:
